Fix bug in undoing ActionRestoreEnvTraced
When restoring an environment, the value stack is popped (the topmost value, the environment, is popped) and the current environment is replaced by this topmost environment.
Undoing this action should therefore not only push back this topmost environment on the stack, as was already being done until now, but it should also make sure the environment that was in use at the moment the ActionRestoreEnv was handled is reinstated. The ActionSingleSaveEnv therefor now consists of an environment (envToSave) that must be pushed back onto the value stack and an envToReplace that must replace the environment.