Fixed bug in running abstract interpretation of program
When handling an ActionStepInTraced while in abstract mode, the machine didn't take into account that the value to be applied as the operator could be something other than a closure. In that case Semantics.bindClosureArgs would return an empty set. The ConcreteTracingProgramState assumed this set was non-empty (as indeed it should be) and AbstractTracingProgramState simply re-used this method