Add ActionDefineAddressesPopR to be used when defining a new address, followed by popping the topmost continuation frame
Most instances of ActionDefineAddressesR (actually, all instances except for the evaluation of a SchemeDefineFunction expression) had to retrieve the saved values from the topmost continuation frame. This action was then followed by popping that frame.
However, retrieving the frame and actually popping it was not performed by the ActionDefineAddressesR action, but was split over the ActionDefineAddressesR and a subsequent ActionPopKontT, which causes a mismatch between the states generated by applying these actions, and the filter edge for these states. In other words, some of the filter edges generated actually could not arise and should be removed