When applying a FrameGenerator, also pass current top frame
If not, all prior values that were saved (e.g., prior operand values during the propagation phase) in this top frame will be lost.
Before this commit, when evaluating an expression where more than one value needs to be saved, the AAM would generate FrameGenerators for each step in this evaluation.
These saved FrameGenerators would still contain values computed during the initial analysis though, so there was no proper separation of data flow.
This issue is solved by passing the current top-frame during the propagation phase so that the ActionRApplier can extract the values saved in that frame and combine them with the new value that was computed for this continuation.