Quickfix for bug
When a RT analysis is performed, the current path is reset.
However, the latest constraint (the one that triggered the RT analysis in the first place) was already added to that path.
As the partial matcher produced by the analysis has not yet followed a branch of this constraint, the constraint should be re-added it can successfully match.